requirements.txt
awsebcli==3.15.3
bootstrap4==0.1.0
boto3==1.10.9
botocore==1.13.16
cement==2.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.3.9
Django==2.0
django-crispy-forms==1.8.0
django-jquery==3.1.0
django-session-security==2.6.5
django-storages==1.7.2
docutils==0.15.2
elasticsearch==7.0.5
et-xmlfile==1.0.1
future==0.16.0
idna==2.7
jdcal==1.4.1
jmespath==0.9.4
joblib==0.14.0
nltk==3.4.5
numpy==1.17.3
openpyxl==3.0.0
pandas==0.25.3
pathspec==0.5.9
Pillow==7.1.2
pyad==0.6.0
python-dateutil==2.8.0
PyYAML==3.13
requests==2.20.1
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
semantic-version==2.5.0
simpletransformers==0.26.0
six==1.11.0
sklearn==0.0
sqlparse==0.3.0
termcolor==1.1.0
troposphere==2.6.0
transformers==2.8.0
urllib3==1.24.3
virtualenv==16.7.7
wcwidth==0.1.7
Werkzeug==0.16.1
wsgi-request-logger==0.4.6